client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.60.0",
    "mcp==1.12.3",
    "python-dotenv==1.1.1",
]
